+++ sentry-ruby/lib/sentry/utils/http_tracing.rb
@@ -19,7 +19,7 @@ module Sentry
         crumb = Sentry::Breadcrumb.new(
           level: :info,
           category: self.class::BREADCRUMB_CATEGORY,
-          type: :info,
+          type: "info",
           data: { status: response_status, **request_info }
         )
 
@@ -35,6 +35,25 @@ module Sentry
           Sentry.initialized? &&
           Sentry.configuration.propagate_traces &&
           Sentry.configuration.trace_propagation_targets.any? { |target| url.match?(target) }
+      end
+
+      # Kindly borrowed from Rack::Utils
+      def build_nested_query(value, prefix = nil)
+        case value
+        when Array
+          value.map { |v|
+            build_nested_query(v, "#{prefix}[]")
+          }.join("&")
+        when Hash
+          value.map { |k, v|
+            build_nested_query(v, prefix ? "#{prefix}[#{k}]" : k)
+          }.delete_if(&:empty?).join("&")
+        when nil
+          URI.encode_www_form_component(prefix)
+        else
+          raise ArgumentError, "value must be a Hash" if prefix.nil?
+          "#{URI.encode_www_form_component(prefix)}=#{URI.encode_www_form_component(value)}"
+        end
       end
     end
   end